feat: use update lock file after dependencies update
