+++ lib/npm-utils.js
@@ -18,6 +18,7 @@ const fs = require('fs');
 const path = require('path');
 const exec = require('child_process').exec;
 const { handleError } = require('./git-utils');
+const { processLockFile } = require('./lock-file-utils');
 
 /**
  * Changes package.json version using npm
@@ -59,7 +60,7 @@ function changeLernaProjectVersion(version, branchName = null) {
  * @param {Object} packageJson - The package.json object
  * @param {Function} versionPredicate - Function that returns true if version should be updated
  * @param {string} newVersion - The new version to set
- * @returns {Array} List of updated packages
+ * @returns {Array} Array of updated packages
  */
 function updateDistTagDependencies(packageJson, versionPredicate, newVersion) {
     const dependencyTypes = ['dependencies', 'devDependencies', 'peerDependencies'];
@@ -148,6 +149,25 @@ function updateLernaPackagesDistTagDependencies(versionPredicate, newVersion) {
 }
 
 /**
+ * Extracts scopes from a list of package names
+ * 
+ * @param {Array} packages - Array of package names
+ * @returns {Array} Array of unique scopes
+ */
+function extractScopes(packages) {
+    const scopes = new Set();
+    
+    packages.forEach(pkg => {
+        if (pkg.startsWith('@')) {
+            const scopePart = pkg.split('/')[0];
+            scopes.add(scopePart);
+        }
+    });
+    
+    return [...scopes];
+}
+
+/**
  * Updates dist tags dependencies and lock files
  * 
  * @param {boolean} isLernaProject - Whether this is a lerna project
@@ -170,14 +190,29 @@ function updateDistTagsDependenciesAndLockFiles(isLernaProject, versionPredicate
         Promise.all(updatePromises).then((results) => {
             // Flatten and combine all updated packages
             const updatedPackages = results.flat();
-            console.log("Updated packages: ", updatedPackages);
-            if (updatedPackages.length > 0) {
-                const uniquePackages = [...new Set(updatedPackages)];
-                exec(`npm update ${uniquePackages.join(' ')}`, (err) => {
-                    if (err) throw err;
-                    console.log("Updated lock files");
+            const uniquePackages = [...new Set(updatedPackages)];
+            
+            console.log("Updated packages:", uniquePackages);
+            
+            if (uniquePackages.length > 0) {
+                // Extract scopes from the updated packages
+                const scopes = extractScopes(uniquePackages);
+                console.log("Affected scopes:", scopes);
+                
+                if (scopes.length > 0) {
+                    // Use processLockFile from lock-file-utils to update lock files for the affected scopes
+                    processLockFile(scopes)
+                        .then(() => {
+                            console.log("Updated lock files for scopes:", scopes);
+                            resolve();
+                        })
+                        .catch(err => {
+                            handleError(err);
+                            resolve();
+                        });
+                } else {
                     resolve();
-                });
+                }
             } else {
                 resolve();
             }
